refactor after conversations
