+++ frontend/src/pages/NotFound/index.ts
@@ -1,1 +1,1 @@
-export { default } from './NotFoundContainer';
+export { default } from './NotFound';
